WellspaceAI/requirements.txt
annotated-types==0.7.0
blinker
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click
colorama
contourpy==1.3.0
cycler==0.12.1
Flask
fonttools==4.53.1
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib-metadata
importlib_resources==6.4.5
itsdangerous
Jinja2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug
zipp
